requirements.txt
absl-py==2.1.0
astunparse==1.6.3
blinker==1.9.0
build==1.2.2.post1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
Flask==2.3.2
Flask-Cors==3.0.10
flatbuffers==25.1.24
gast==0.6.0
google-pasta==0.2.0
grpcio==1.70.0
h5py==3.12.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
keras==3.8.0
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.4.1
namex==0.0.8
numpy==2.2.2
opt_einsum==3.4.0
optree==0.14.0
packaging==24.2
Pillow==10.0.0
pip-tools==7.4.1
protobuf==5.29.3
Pygments==2.19.1
pyproject_hooks==1.2.0
requests==2.32.3
rich==13.9.4
scikit-learn==1.6.1
scipy==1.15.1
setuptools==75.8.0
six==1.17.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
termcolor==2.5.0
threadpoolctl==3.5.0
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
